fix: shared type 수정/ 관련된  경로 수정
