+++ src/shared/types/index.ts
@@ -1,1 +1,5 @@
 export * from './item';
+export * from './feed';
+export * from './hobby';
+export * from './member';
+export * from './review';
