+++ src/shared/types/member.ts
@@ -1,0 +1,6 @@
+export interface MemberInfo {
+  memberId: number;
+  nickName: string;
+  profileImage: string;
+  level: number;
+}
