+++ src/shared/types/review.ts
@@ -1,0 +1,11 @@
+import { MemberInfo } from './member';
+
+export interface Review {
+  cursorId: string;
+  memberInfo: MemberInfo;
+  reviewId: number;
+  rate: number;
+  content: string;
+  createdAt: string;
+  updatedAt: string;
+}
